requirements.txt
google-genai
pydantic
python-dotenv
python-chess
